requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# LLM integrations
openai>=1.0.0
google-generativeai>=0.3.0

# API framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
celery>=5.3.0
redis>=5.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
tenacity>=8.2.0

# Monitoring and logging
prometheus-client>=0.18.0
structlog>=23.2.0

# Data visualization (for analysis)
matplotlib>=3.8.0
seaborn>=0.13.0

# Development tools
black>=23.10.0
ruff>=0.1.0
mypy>=1.6.0
